add more test cases for loader
