+++ cmd/packageloader/loader/loader_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/skycoin/cx/cmd/packageloader/loader"
-	"github.com/skycoin/cx/cx/ast"
 )
 
 const TEST_SRC_PATH_VALID = "test_folder/test_invalid_program/src/"
@@ -37,7 +36,7 @@ func TestGetPackageName(t *testing.T) {
 }
 
 func TestCreateFileMap(t *testing.T) {
-	_, sourceCodes, _ := ast.ParseArgsForCX([]string{TEST_SRC_PATH_VALID}, true)
+	_, sourceCodes, _ := loader.ParseArgsForCX([]string{TEST_SRC_PATH_VALID}, true)
 	_, err := loader.CreateFileMap(sourceCodes)
 	if err != nil {
 		t.Error(err)
@@ -56,5 +55,68 @@ func TestFileStructFromFile(t *testing.T) {
 	}
 	if testFileStruct.FileName != "package.cx" {
 		t.Error("wrong file name:", testFileStruct.FileName)
+	}
+}
+
+func TestAddNewPackage(t *testing.T) {
+	tests := []struct {
+		Scenario  string
+		FilesPath string
+		Database  string
+	}{
+		{
+			Scenario:  "Test adding package to Redis database",
+			FilesPath: TEST_SRC_PATH_VALID,
+			Database:  "redis",
+		},
+		{
+			Scenario:  "Test adding package to Bolt database",
+			FilesPath: TEST_SRC_PATH_VALID,
+			Database:  "bolt",
+		},
+	}
+	for _, testcase := range tests {
+		t.Run(testcase.Scenario, func(t *testing.T) {
+			_, sourceCodes, _ := loader.ParseArgsForCX([]string{testcase.FilesPath}, false)
+			testPackageStruct := loader.PackageList{}
+			err := loader.AddNewPackage(&testPackageStruct, "main", sourceCodes, testcase.Database)
+			if err != nil {
+				t.Error(err)
+			}
+			if len(testPackageStruct.Packages) != 1 {
+				t.Error("Wrong number of packages added")
+			}
+		})
+	}
+}
+
+func TestLoadCXProgram(t *testing.T) {
+	tests := []struct {
+		Scenario         string
+		FilesPath        string
+		Database         string
+		ExpectedPackages int
+	}{
+		{
+			Scenario:         "Test adding package to Redis database",
+			FilesPath:        TEST_SRC_PATH_VALID,
+			Database:         "redis",
+			ExpectedPackages: 2,
+		},
+		{
+			Scenario:         "Test adding package to Bolt database",
+			FilesPath:        TEST_SRC_PATH_VALID,
+			Database:         "bolt",
+			ExpectedPackages: 2,
+		},
+	}
+	for _, testcase := range tests {
+		t.Run(testcase.Scenario, func(t *testing.T) {
+			_, sourceCodes, _ := loader.ParseArgsForCX([]string{testcase.FilesPath}, false)
+			err := loader.LoadCXProgram("test", sourceCodes, testcase.Database)
+			if err != nil {
+				t.Error(err)
+			}
+		})
 	}
 }
